Move logger into candlelib
