+++ candlelib/include/candle/MD/MD.hpp
@@ -2,7 +2,7 @@
 
 #include "candle/types/mab_types.hpp"
 #include "candle/types/md_types.hpp"
-#include "logger.hpp"
+#include "candle/logger/logger.hpp"
 #include "candle/types/manufacturer_data.hpp"
 #include "candle/communication_device/candle_types.hpp"
 #include "candle/MD/MDStatus.hpp"
