+++ candlelib/include/candle/communication_device/candle_frame_adapter.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "logger.hpp"
+#include "candle/logger/logger.hpp"
 #include "candle/types/candle_frame_dto.hpp"
 #include "candle/types/mab_types.hpp"
