+++ candlelib/include/candle/communication_interface/SPI.hpp
@@ -13,7 +13,7 @@
 
 #include "candle/types/mab_types.hpp"
 #include "candle/communication_interface/I_communication_interface.hpp"
-#include "logger.hpp"
+#include "candle/logger/logger.hpp"
 #include "candle/crc.hpp"
 
 namespace mab
