+++ candlelib/include/candle/pds/pds.hpp
@@ -3,7 +3,7 @@
 #include <stdint.h>
 
 #include "candle/communication_device/candle.hpp"
-#include "logger.hpp"
+#include "candle/logger/logger.hpp"
 
 #include "candle/types/mab_types.hpp"
 #include "candle/pds/pds_module.hpp"
