+++ candlelib/src/communication_interface/SPI.cpp
@@ -7,7 +7,7 @@
 #include <sys/ioctl.h>
 
 #include "candle/communication_interface/I_communication_interface.hpp"
-#include "logger.hpp"
+#include "candle/logger/logger.hpp"
 #include "candle/communication_interface/SPI.hpp"
 
 namespace mab
