+++ candlelib/test/communication_device/candle_bootloader_test.cpp
@@ -1,7 +1,7 @@
 #include "candle/communication_interface/I_communication_interface_mock.hpp"
 
 #include "candle/types/mab_types.hpp"
-#include "logger.hpp"
+#include "candle/logger/logger.hpp"
 #include "candle/communication_device/candle_bootloader.hpp"
 
 #include <gtest/gtest.h>
