+++ examples/cpp/app_template.cpp
@@ -1,7 +1,7 @@
 #include "candle/communication_device/candle.hpp"
 #include "candle/pds/pds.hpp"
 #include "candle/MD/MD.hpp"
-#include "logger.hpp"
+#include "candle/logger/logger.hpp"
 
 #include <vector>
 #include <thread>
